// ClientApp/src/store/interface/interface.types.ts
export enum INTERFACE_ACTION_TYPES  {
    CREATE_START = 'interface/CREATE_START',
    CREATE_SUCCESS = 'interface/CREATE_SUCCESS',
    CREATE_FAILED = 'interface/CREATE_FAILED',
    UPDATE_START = 'interface/UPDATE_START',
    UPDATE_SUCCESS = 'interface/UPDATE_SUCCESS',
    UPDATE_FAILED = 'interface/UPDATE_FAILED',
    DELETE_START = 'interface/DELETE_START',
    DELETE_SUCCESS = 'interface/DELETE_SUCCESS',
    DELETE_FAILED = 'interface/DELETE_FAILED',
    FETCH_SINGLE_START = 'interface/FETCH_SINGLE_START',
    FETCH_SINGLE_SUCCESS = 'interface/FETCH_SINGLE_SUCCESS',
    FETCH_SINGLE_FAILED = 'interface/FETCH_SINGLE_FAILED',
    FETCH_USER_INTERFACE_START = 'interface/FETCH_USER_INTERFACE_START',
    FETCH_USER_INTERFACE_SUCCESS = 'interface/FETCH_USER_INTERFACE_SUCCESS',
    FETCH_USER_INTERFACE_FAILED = 'interface/FETCH_USER_INTERFACE_FAILED',
    FETCH_OTHER_USER_INTERFACE_START = 'interface/FETCH_OTHER_USER_INTERFACE_START',
    FETCH_OTHER_USER_INTERFACE_SUCCESS = 'interface/FETCH_OTHER_USER_INTERFACE_SUCCESS',
    FETCH_OTHER_USER_INTERFACE_FAILED = 'interface/FETCH_OTHER_USER_INTERFACE_FAILED',
    FETCH_ALL_START = 'interface/FETCH_ALL_START',
    FETCH_ALL_SUCCESS = 'interface/FETCH_ALL_SUCCESS',
    FETCH_ALL_FAILED = 'interface/FETCH_ALL_FAILED',
    TOGGLE_UTILS = 'interface/TOGGLE_UTILS'
};

export type Interface = {
    utilsOpen: boolean;
}